pyproject.toml
[tool.poetry]
name = "polymind"
version = "0.0.26"  # Update this version before publishing to PyPI
description = "PolyMind is a customizable collaborative multi-agent framework for collective intelligence and distributed problem solving."
authors = ["TechTao"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
numpy = "1.26.0"
qdrant-client = "1.7.0"
pydantic = "^2.6.3"
python-dotenv = "^1.0.1"
openai = "^1.14.0"
aiohttp = "^3.9.3"
pymilvus = "2.3.7"
faiss-cpu = "1.8.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
aioresponses = "^0.7.6"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"

[tool.black]
line-length = 120

[tool.isort]
skip = ["__init__.py"]

[tool.flake8]
ignore = "Q000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
